@@ -1,6 +1,6 @@
 /**
  * Created By shoh@n
- * Date: 8/6/2022
+ * Date: 1/16/2023
  */
 
 package com.example.demo.configs;
@@ -8,9 +8,16 @@ package com.example.demo.configs;
 import org.modelmapper.ModelMapper;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
+import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
+import org.springframework.security.crypto.password.PasswordEncoder;
 
 @Configuration
-public class ModelMapperConfig {
+public class AppConfigs {
+
+    @Bean
+    public PasswordEncoder passwordEncoder() {
+        return new BCryptPasswordEncoder(11);
+    }
 
     @Bean
     ModelMapper modelMapper() {
